loot_sweep: decimal points (2dp), 'also counts as' aliases, required-count gating
- Receipts pay 2-decimal points (a 1-pointer's second receipt at 20% decay
  is 0.8, not 1): receipt_points/item_points round per receipt AND per
  aggregate; engine apply/revoke deltas round at 2dp; web_event_teams.score
  + web_event_progress.progress migrated int -> DOUBLE (applied to prod);
  score_num serializes int-when-integral so other kinds stay clean.
- Items take match_names (alternate drop names crediting the same entry's
  counter/decay/cap — vestige + gold ring) registered in the matcher index
  and pooled at fold time, plus required N (receipts, any mix of names,
  before the entry counts toward group completion — any 3 ancestral
  pieces; repeats every N for bonus_max > 1). Validator snaps aliases to
  canonical names and reserves them task-wide; board + receipts endpoints
  expose required/match_names/matched_name.

+++ db/models/events.py
@@ -383,7 +383,10 @@ class EventTeam(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
     event_id = Column(Integer, ForeignKey("web_events.id"), nullable=False)
     name = Column(String(80), nullable=False)
-    score = Column(Integer, nullable=False, default=0)
+    # DOUBLE (2026-07-19): loot_sweep awards decimal points (a 1-pointer's
+    # second receipt at 20% decay is 0.8). Every write rounds to 2dp; other
+    # event kinds keep integral values.
+    score = Column(Float(53), nullable=False, default=0)
     # The clan this team represents (clan_vs_clan only; NULL on standard/global).
     group_id = Column(Integer, ForeignKey("groups.group_id"), nullable=True)
     # Admin-assigned accent color ("#rrggbb"); NULL = frontend palette default.
@@ -631,8 +634,10 @@ class EventProgress(Base):
     event_id = Column(Integer, ForeignKey("web_events.id"), nullable=False)
     task_id = Column(Integer, ForeignKey("web_event_tasks.id"), nullable=False)
     team_id = Column(Integer, ForeignKey("web_event_teams.id"), nullable=False)
-    # BigInteger (P1-7): loot_value progress folds raw GP; see EventCompletion.
-    progress = Column(BigInteger, nullable=False, default=0)
+    # DOUBLE (2026-07-19, was BigInteger): loot_value progress folds raw GP
+    # (integers stay exact to 2^53 ≈ 9e15) and loot_sweep running totals are
+    # 2-decimal point values.
+    progress = Column(Float(53), nullable=False, default=0)
     completed = Column(Boolean, nullable=False, default=False)
     completed_at = Column(DateTime, nullable=True)
